github comment fixes
