+++ apps/server/src/routes/agent/index.ts
@@ -48,8 +48,6 @@ import { connectionToDriver, getZeroSocketAgent, reSyncThread } from '../../lib/
 import { generateWhatUserCaresAbout, type UserTopic } from '../../lib/analyze/interests';
 import { DurableObjectOAuthClientProvider } from 'agents/mcp/do-oauth-client-provider';
 import { AiChatPrompt, GmailSearchAssistantSystemPrompt } from '../../lib/prompts';
-import { connection, subscriptions, subscriptionThreads } from '../../db/schema';
-import { and, desc, eq, isNotNull, sql } from 'drizzle-orm';
 import { Migratable, Queryable, Transfer } from 'dormroom';
 import type { CreateDraftData } from '../../lib/schemas';
 import { drizzle } from 'drizzle-orm/durable-sqlite';
@@ -57,13 +55,15 @@ import { getPrompt } from '../../pipelines.effect';
 import { AIChatAgent } from 'agents/ai-chat-agent';
 import { DurableObject } from 'cloudflare:workers';
 import { ToolOrchestrator } from './orchestrator';
+import { desc, eq, isNotNull } from 'drizzle-orm';
 import migrations from './db/drizzle/migrations';
 import { getPromptName } from '../../pipelines';
 import { anthropic } from '@ai-sdk/anthropic';
-import { env, type ZeroEnv } from '../../env';
+import { connection } from '../../db/schema';
 import type { WSMessage } from 'partyserver';
 import { tools as authTools } from './tools';
 import { processToolCalls } from './utils';
+import { type ZeroEnv } from '../../env';
 import { type Connection } from 'agents';
 import { openai } from '@ai-sdk/openai';
 import * as schema from './db/schema';
@@ -856,279 +856,6 @@ export class ZeroDriver extends DurableObject<ZeroEnv> {
       throw new Error('No driver available');
     }
     return await this.driver.deleteLabel(id);
-  }
-
-  // Subscription Management Methods
-  async listSubscriptions(params: {
-    userId: string;
-    connectionId?: string;
-    category?: string;
-    isActive?: boolean;
-    limit?: number;
-    offset?: number;
-  }) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      const conditions = [eq(subscriptions.userId, params.userId)];
-
-      if (params.connectionId) {
-        conditions.push(eq(subscriptions.connectionId, params.connectionId));
-      }
-
-      if (params.category) {
-        conditions.push(eq(subscriptions.category, params.category));
-      }
-
-      if (params.isActive !== undefined) {
-        conditions.push(eq(subscriptions.isActive, params.isActive));
-      }
-
-      const [items, totalResult] = await Promise.all([
-        db
-          .select({
-            id: subscriptions.id,
-            senderEmail: subscriptions.senderEmail,
-            senderName: subscriptions.senderName,
-            senderDomain: subscriptions.senderDomain,
-            category: subscriptions.category,
-            listUnsubscribeUrl: subscriptions.listUnsubscribeUrl,
-            listUnsubscribePost: subscriptions.listUnsubscribePost,
-            lastEmailReceivedAt: subscriptions.lastEmailReceivedAt,
-            emailCount: subscriptions.emailCount,
-            isActive: subscriptions.isActive,
-            userUnsubscribedAt: subscriptions.userUnsubscribedAt,
-            autoArchive: subscriptions.autoArchive,
-            metadata: subscriptions.metadata,
-            createdAt: subscriptions.createdAt,
-          })
-          .from(subscriptions)
-          .where(and(...conditions))
-          .orderBy(desc(subscriptions.lastEmailReceivedAt))
-          .limit(params.limit || 50)
-          .offset(params.offset || 0),
-        db
-          .select({ count: sql<number>`count(*)` })
-          .from(subscriptions)
-          .where(and(...conditions)),
-      ]);
-
-      const total = totalResult[0]?.count || 0;
-
-      return {
-        items,
-        total,
-        hasMore: (params.offset || 0) + items.length < total,
-      };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async getSubscription(subscriptionId: string, userId: string) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      const [subscription] = await db
-        .select()
-        .from(subscriptions)
-        .where(and(eq(subscriptions.id, subscriptionId), eq(subscriptions.userId, userId)));
-
-      if (!subscription) {
-        throw new Error('Subscription not found');
-      }
-
-      // Get recent threads
-      const recentThreads = await db
-        .select({
-          threadId: subscriptionThreads.threadId,
-          messageId: subscriptionThreads.messageId,
-          receivedAt: subscriptionThreads.receivedAt,
-          subject: subscriptionThreads.subject,
-        })
-        .from(subscriptionThreads)
-        .where(eq(subscriptionThreads.subscriptionId, subscriptionId))
-        .orderBy(desc(subscriptionThreads.receivedAt))
-        .limit(10);
-
-      return {
-        ...subscription,
-        recentThreads,
-      };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async unsubscribeFromEmail(subscriptionId: string, userId: string) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      // Get subscription details
-      const [subscription] = await db
-        .select()
-        .from(subscriptions)
-        .where(and(eq(subscriptions.id, subscriptionId), eq(subscriptions.userId, userId)));
-
-      if (!subscription) {
-        throw new Error('Subscription not found');
-      }
-
-      // Update subscription as inactive
-      await db
-        .update(subscriptions)
-        .set({
-          isActive: false,
-          userUnsubscribedAt: new Date(),
-          updatedAt: new Date(),
-        })
-        .where(eq(subscriptions.id, subscriptionId));
-
-      // If there's a List-Unsubscribe header, return the action
-      let unsubscribeAction = null;
-      if (subscription.listUnsubscribeUrl) {
-        const { getListUnsubscribeAction } = await import('../../lib/email-utils');
-        unsubscribeAction = getListUnsubscribeAction({
-          listUnsubscribe: subscription.listUnsubscribeUrl,
-          listUnsubscribePost: subscription.listUnsubscribePost || undefined,
-        });
-      }
-
-      return {
-        success: true,
-        unsubscribeAction,
-      };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async resubscribeToEmail(subscriptionId: string, userId: string) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      await db
-        .update(subscriptions)
-        .set({
-          isActive: true,
-          userUnsubscribedAt: null,
-          updatedAt: new Date(),
-        })
-        .where(and(eq(subscriptions.id, subscriptionId), eq(subscriptions.userId, userId)));
-
-      return { success: true };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async updateSubscriptionPreferences(params: {
-    subscriptionId: string;
-    userId: string;
-    autoArchive?: boolean;
-    category?: string;
-  }) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      const updateData: any = {
-        updatedAt: new Date(),
-      };
-
-      if (params.autoArchive !== undefined) {
-        updateData.autoArchive = params.autoArchive;
-      }
-
-      if (params.category) {
-        updateData.category = params.category;
-      }
-
-      await db
-        .update(subscriptions)
-        .set(updateData)
-        .where(
-          and(eq(subscriptions.id, params.subscriptionId), eq(subscriptions.userId, params.userId)),
-        );
-
-      return { success: true };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async bulkUnsubscribeEmails(subscriptionIds: string[], userId: string) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      await db
-        .update(subscriptions)
-        .set({
-          isActive: false,
-          userUnsubscribedAt: new Date(),
-          updatedAt: new Date(),
-        })
-        .where(
-          and(eq(subscriptions.userId, userId), sql`${subscriptions.id} = ANY(${subscriptionIds})`),
-        );
-
-      return { success: true, count: subscriptionIds.length };
-    } finally {
-      await conn.end();
-    }
-  }
-
-  async getSubscriptionStats(userId: string, connectionId?: string) {
-    const { db, conn } = createDb(env.HYPERDRIVE.connectionString);
-
-    try {
-      const conditions = [eq(subscriptions.userId, userId)];
-
-      if (connectionId) {
-        conditions.push(eq(subscriptions.connectionId, connectionId));
-      }
-
-      // Get stats by category
-      const categoryStats = await db
-        .select({
-          category: subscriptions.category,
-          count: sql<number>`count(*)`,
-          activeCount: sql<number>`count(*) filter (where ${subscriptions.isActive} = true)`,
-        })
-        .from(subscriptions)
-        .where(and(...conditions))
-        .groupBy(subscriptions.category);
-
-      // Get overall stats
-      const [overallStats] = await db
-        .select({
-          total: sql<number>`count(*)`,
-          active: sql<number>`count(*) filter (where ${subscriptions.isActive} = true)`,
-          inactive: sql<number>`count(*) filter (where ${subscriptions.isActive} = false)`,
-          avgEmailsPerSubscription: sql<number>`avg(${subscriptions.emailCount})`,
-        })
-        .from(subscriptions)
-        .where(and(...conditions));
-
-      // Get recent activity
-      const thirtyDaysAgo = new Date();
-      thirtyDaysAgo.setDate(thirtyDaysAgo.getDate() - 30);
-
-      const [recentActivity] = await db
-        .select({
-          recentlyReceived: sql<number>`count(*) filter (where ${subscriptions.lastEmailReceivedAt} >= ${thirtyDaysAgo})`,
-          recentlyUnsubscribed: sql<number>`count(*) filter (where ${subscriptions.userUnsubscribedAt} >= ${thirtyDaysAgo})`,
-        })
-        .from(subscriptions)
-        .where(and(...conditions));
-
-      return {
-        overall: overallStats,
-        byCategory: categoryStats,
-        recentActivity,
-      };
-    } finally {
-      await conn.end();
-    }
   }
 
   async createDraft(draftData: CreateDraftData) {
